rename abstract command helper class
